Feat: RL1M-304 웹소켓 페이로드 타입 추론 가능하게 - request schema에서 request mapper를 optional하게 정의할 수 있게
